Fixed Anisotropic filter init

+++ openb3dlib.mod/openb3d/src/texture.h
@@ -48,7 +48,7 @@ public:
 	float multitex_factor;
 	static int isunique;
 	int format; // can be GL_RGBA or one of the DXT compressed ones
-	static float AnIsoSupport;
+	static int AnIsoSupport;
 	
 	Texture(){
